Add more entities:
 - DeviceScript
 - GlobalSettings (singleton)
 - CriteriaScript
 - EntitlementScript

NOTE: DeviceScript has `readOnly` fields that we can get but not post. They are
not supported now but they should. Those fields are checksums, we should get
them but make them Optional. If there we should use to check the checksum at
some point.

It also has a `writeOnly` field. This is even more difficult to integrate. It's
used on POST (base64 bytes for the file) but not on get. We should do something
like this:

 - link the files with files on a directory (or uploaded into k8s) and use the
 checksum field to compare them.
 - create the base64 only when doing POST.