Add env variable `APPGATE_OPERATOR_DUMP_SECRETS`.
When set this variables forces to dump the `writeOnly` attributes with format
`password` always.

This means that (since the controller never returns those fields) when comparing
we will always PUT those entities.

TODO: Maybe we want to run it only once when this flag is on so the secrets are
upadted and then we run in normal mode after.