Document the new APPGATE_OPERATOR_*_TAGS environment variables
We have 3 kind of tags:

An entity belogs to a tag set if it has tags that are members of the set.

- BUILTIN_TAGS

Those are the entities considered builtin in the system. They will never be
deleted and they will kept when running in clean up mode.

- TARGET_TAGS

When defined these are the entities we are interested on. Tags not member of
this set won't be retrieved from the controller. If empty all the entities
belong to this set.

- FILTER_TAGS

Set of entities that we want to filter out explicitly from the working plan.