Make K8SDumper to dump the whole data needed for kubernetes
Until now the dumper was just dumping the spec and a wrapper was taking
care of the rest. This makes k8s dumper to generate all the data.

This change has been made to be able to serialize easily information
from AppgateMetadata into k8s. Right now only the id is dumped as an
annotation.

Now, when an entity is dumped into k8s it will have an annotation with
the id