Add k8s annotation for fields that contain uuids in references
The git operator will try to resolve all the uuids into names of
entities if possible. If it's not possible it will dump the uuid (id of
the entities referenced).

Later we will need to reload them so we will need to know which ones are
names and which ones uuid